Reorganize tests to share more between envs
This commit restructures the tests (and some code) to allow the
rendering test suite to be written once and used with rehydration, using
the lazy builder, and using the bundle compiler.

This is accomplished by moving the environment specific questions into a
RenderDelegate, which is responsible for:

- constructing the initial element to render into
- registering components
- registering helpers
- rendering a template for a context into an element

At the moment, three rehydration tests fail, but I'm not entirely sure
how they were ever passing. They involve quirks of browser parsing (for
example, <table><!-- comment --><tr></tr><!-- comment --></table> is
parsed as <table><!-- comment --><tbody><tr></tr><!-- comment
--></tbody></table>, which mis-nests rehydration markers.

We also have tests that confirm that <option selected={{value}}> always
produces <option> even when selected is truthy. These tests seem like
they're testing quirks of the current attribute implementation, when SSR
would of course need to retain those attributes.

I might have made a mistake, or these tests might not have been running
correctly before.
