+++ packages/@glimmer/test-helpers/lib/abstract-test-case.ts
@@ -105,20 +105,33 @@ class SimplePathReference implements PathReference<Opaque> {
 type IndividualSnapshot = 'up' | 'down' | Node;
 type NodesSnapshot = IndividualSnapshot[];
 
-export abstract class AbstractRenderTest {
-  protected abstract element: HTMLElement;
-  protected serialized: string;
+export interface RenderDelegate {
+  getInitialElement(): HTMLElement;
+  registerComponent<K extends ComponentKind>(type: K, name: string, layout: string, Class?: ComponentTypes[K]): void;
+  registerHelper(name: string, helper: UserHelper): void;
+  renderTemplate(template: string, context: Dict<Opaque>, element: HTMLElement, snapshot: () => void): RenderResult;
+}
+
+export class AbstractRenderTest {
+  protected element: HTMLElement;
   protected assert = QUnit.assert;
   protected context: Dict<Opaque> = dict<Opaque>();
   protected renderResult: Option<RenderResult> = null;
   protected helpers = dict<UserHelper>();
+  protected testType: ComponentKind;
   private snapshot: NodesSnapshot = [];
-  public testType: ComponentKind;
-  public template: string;
 
-  abstract registerComponent(type: ComponentKind, name: string, layout: string): void;
-  abstract registerHelper(name: string, helper: UserHelper): void;
-  // abstract populateHelpers(): void;
+  constructor(protected delegate: RenderDelegate) {
+    this.element = delegate.getInitialElement();
+  }
+
+  registerHelper(name: string, helper: UserHelper) {
+    this.delegate.registerHelper(name, helper);
+  }
+
+  registerComponent<K extends ComponentKind>(type: K, name: string, layout: string, Class?: ComponentTypes[K]) {
+    this.delegate.registerComponent(type, name, layout, Class);
+  }
 
   buildComponent(blueprint: ComponentBlueprint): string {
     let invocation = '';
@@ -234,7 +247,7 @@ export abstract class AbstractRenderTest {
     let invocation = this.buildAngleBracketComponent(blueprint);
     let layoutAttrs = this.buildAttributes(blueprint.layoutAttributes);
     this.assert.ok(true, `generated glimmer layout as ${`<${tag} ${layoutAttrs} ...attributes>${layout}</${tag}>`}`);
-    this.registerComponent("Glimmer", name, `<${tag} ${layoutAttrs} ...attributes>${layout}</${tag}>`);
+    this.delegate.registerComponent("Glimmer", name, `<${tag} ${layoutAttrs} ...attributes>${layout}</${tag}>`);
     this.assert.ok(true, `generated glimmer invocation as ${invocation}`);
     return invocation;
   }
@@ -284,7 +297,7 @@ export abstract class AbstractRenderTest {
       invocation.push('}}');
     }
     this.assert.ok(true, `generated curly layout as ${layout}`);
-    this.registerComponent("Curly", name, layout);
+    this.delegate.registerComponent("Curly", name, layout);
     invocation = invocation.join("");
     this.assert.ok(true, `generated curly invocation as ${invocation}`);
     return invocation;
@@ -294,7 +307,7 @@ export abstract class AbstractRenderTest {
     let { layout, name = GLIMMER_TEST_COMPONENT } = blueprint;
     let invocation = this.buildAngleBracketComponent(blueprint);
     this.assert.ok(true, `generated fragment layout as ${layout}`);
-    this.registerComponent("Basic", name, `${layout}`);
+    this.delegate.registerComponent("Basic", name, `${layout}`);
     this.assert.ok(true, `generated fragment invocation as ${invocation}`);
     return invocation;
   }
@@ -303,7 +316,7 @@ export abstract class AbstractRenderTest {
     let { tag = "div", layout, name = GLIMMER_TEST_COMPONENT } = blueprint;
     let invocation = this.buildAngleBracketComponent(blueprint);
     this.assert.ok(true, `generated basic layout as ${layout}`);
-    this.registerComponent("Basic", name, `<${tag} ...attributes>${layout}</${tag}>`);
+    this.delegate.registerComponent("Basic", name, `<${tag} ...attributes>${layout}</${tag}>`);
     this.assert.ok(true, `generated basic invocation as ${invocation}`);
     return invocation;
   }
@@ -343,7 +356,7 @@ export abstract class AbstractRenderTest {
     }
 
     this.assert.ok(true, `generated dynamic layout as ${layout}`);
-    this.registerComponent("Curly", name, layout);
+    this.delegate.registerComponent("Curly", name, layout);
     invocation = invocation.join("");
     this.assert.ok(true, `generated dynamic invocation as ${invocation}`);
 
@@ -353,7 +366,7 @@ export abstract class AbstractRenderTest {
   shouldBeVoid(tagName: string) {
     this.element.innerHTML = "";
     let html = "<" + tagName + " data-foo='bar'><p>hello</p>";
-    this.renderTemplate(html);
+    this.delegate.renderTemplate(html, this.context, this.element, () => this.takeSnapshot());
 
     let tag = '<' + tagName + ' data-foo="bar">';
     let closing = '</' + tagName + '>';
@@ -380,10 +393,8 @@ export abstract class AbstractRenderTest {
 
     this.setProperties(properties);
 
-    this.renderResult = this.renderTemplate(template);
+    this.renderResult = this.delegate.renderTemplate(template, this.context, this.element, () => this.takeSnapshot());
   }
-
-  protected abstract renderTemplate(template: string): RenderResult;
 
   rerender(properties: Dict<Opaque> = {}): void {
     this.setProperties(properties);
@@ -443,8 +454,8 @@ export abstract class AbstractRenderTest {
     this.assertStableNodes();
   }
 
-  protected assertHTML(html: string) {
-    equalTokens(this.element, html, html);
+  protected assertHTML(html: string, message?: string) {
+    equalTokens(this.element, html, message ? `${html} (${message})` : html);
     this.takeSnapshot();
   }
 
@@ -481,63 +492,70 @@ export abstract class AbstractRenderTest {
   }
 }
 
-export abstract class AbstractTestEnvironmentRenderTest extends AbstractRenderTest {
-  protected env: TestEnvironment;
+export class TestEnvironmentRenderDelegate implements RenderDelegate {
+  constructor(protected env: TestEnvironment = new TestEnvironment()) {}
 
-  registerComponent(type: ComponentKind = this.testType, name: string, layout: string) {
-    switch (type) {
-      case "Glimmer":
-        this.env.registerEmberishGlimmerComponent(name, EmberishGlimmerComponent, layout);
-        break;
-      case "Curly":
-        this.env.registerEmberishCurlyComponent(name, EmberishCurlyComponent, layout);
-        break;
+  getInitialElement(): HTMLElement {
+    return this.env.getAppendOperations().createElement('div') as HTMLElement;
+  }
 
-      case "Dynamic":
-        this.env.registerEmberishCurlyComponent(name, EmberishCurlyComponent, layout);
-        break;
-      case "Basic":
-      case "Fragment":
-        this.env.registerBasicComponent(name, BasicComponent, layout);
-        break;
-    }
+  registerComponent<K extends ComponentKind>(type: K, name: string, layout: string, Class?: ComponentTypes[K]) {
+    registerComponent(this.env, type, name, layout, Class);
   }
 
   registerHelper(name: string, helper: UserHelper): void {
     this.env.registerHelper(name, helper);
   }
-}
 
-export class RenderTests extends AbstractTestEnvironmentRenderTest {
-  protected element: HTMLDivElement;
-
-  protected env: TestEnvironment = new TestEnvironment();
-
-  constructor() {
-    super();
-    this.element = this.env.getAppendOperations().createElement("div") as HTMLDivElement;
-  }
-
-  renderTemplate(template: string): RenderResult {
+  renderTemplate(template: string, context: Dict<Opaque>, element: HTMLElement): RenderResult {
     let { env } = this;
     return renderTemplate(template, {
       env,
-      self: new UpdatableReference(this.context),
-      cursor: { element: this.element, nextSibling: null },
+      self: new UpdatableReference(context),
+      cursor: { element, nextSibling: null },
       dynamicScope: new TestDynamicScope()
     });
   }
 }
 
-export class RehydrationTests extends RenderTests {
+export const CLASSES = {
+  Glimmer: EmberishGlimmerComponent,
+  Curly: EmberishCurlyComponent,
+  Dynamic: EmberishCurlyComponent,
+  Basic: BasicComponent,
+  Fragment: BasicComponent
+}
+
+export type ComponentTypes = typeof CLASSES;
+
+function registerComponent<K extends ComponentKind>(env: TestEnvironment, type: K, name: string, layout: string, Class?: ComponentTypes[K]) {
+  switch (type) {
+    case "Glimmer":
+      env.registerEmberishGlimmerComponent(name, Class as typeof EmberishGlimmerComponent, layout);
+      break;
+    case "Curly":
+      env.registerEmberishCurlyComponent(name, Class as typeof EmberishCurlyComponent, layout);
+      break;
+
+    case "Dynamic":
+      env.registerEmberishCurlyComponent(name, Class as typeof EmberishCurlyComponent, layout);
+      break;
+    case "Basic":
+    case "Fragment":
+      env.registerBasicComponent(name, Class as typeof BasicComponent, layout);
+      break;
+  }
+}
+
+export class RehydrationDelegate implements RenderDelegate {
   serialized: string;
 
   protected env: never;
-  protected clientEnv: TestEnvironment;
-  protected serverEnv: TestEnvironment;
+
+  public clientEnv: TestEnvironment;
+  public serverEnv: TestEnvironment;
 
   constructor() {
-    super();
     this.clientEnv = new TestEnvironment();
 
     let doc = new SimpleDOM.Document();
@@ -548,85 +566,58 @@ export class RehydrationTests extends RenderTests {
     });
   }
 
-  setupServer(template: string = this.template) {
-    this.setup({ template });
+  getInitialElement(): HTMLElement {
+    return this.clientEnv.getAppendOperations().createElement('div') as HTMLElement;
   }
 
-  setupClient(template: string = this.template) {
-    let div = document.createElement("div");
-
-    expect(
-      this.serialized,
-      'Should have serialized HTML from `this.renderServerSide()`'
-    );
-
-    div.innerHTML = this.serialized;
-    this.element = div;
-    this.setup({ template });
-  }
-
-  setup({ template, context }: { template: string; context?: Dict<Opaque>; env?: TestEnvironment }) {
-    this.template = template;
-    if (context) this.setProperties(context);
-  }
-
-  assertServerOutput(..._expected: Content[]) {
-    let serialized = this.serialize();
-    equalTokens(serialized, content([OPEN, ..._expected, CLOSE]));
-    this.serialized = serialized;
-  }
-
-  renderServerSide(context?: Dict<Opaque>): void {
-    if (context) {
-      this.context = context;
-    }
-    this.setupServer();
+  renderServerSide(template: string, context: Dict<Opaque>, takeSnapshot: () => void): string {
     let env = this.serverEnv;
-    this.element = env.getAppendOperations().createElement("div") as HTMLDivElement;
-    let template = expect(this.template, "Must set up a template before calling renderServerSide");
+    let element = env.getAppendOperations().createElement("div") as HTMLDivElement;
     // Emulate server-side render
     renderTemplate(template, {
       env,
-      self: new UpdatableReference(this.context),
-      cursor: { element: this.element, nextSibling: null },
+      self: new UpdatableReference(context),
+      cursor: { element, nextSibling: null },
       dynamicScope: new TestDynamicScope(),
       mode: 'serialize'
     });
 
-    this.takeSnapshot();
-    this.serialized = this.serialize();
+    takeSnapshot();
+    return this.serialize(element);
   }
 
-  serialize() {
+  serialize(element: HTMLElement) {
     let serializer = new SimpleDOM.HTMLSerializer(SimpleDOM.voidMap);
-    let serialized = serializer.serializeChildren(this.element);
+    let serialized = serializer.serializeChildren(element);
     return serialized;
   }
 
-  renderClientSide(context?: Dict<Opaque>) {
-    if (context) {
-      this.context = context;
-    }
-    this.setupClient();
-    // this.populateHelpers();
+  renderClientSide(template: string, context: Dict<Opaque>, element: HTMLElement): RenderResult {
     let env = this.clientEnv;
-    this.element = env.getAppendOperations().createElement("div") as HTMLDivElement;
-    let template = expect(this.template, "Must set up a template before calling renderClientSide");
     // Client-side rehydration
-    this.renderResult = renderTemplate(template, {
+    return renderTemplate(template, {
       env,
-      self: new UpdatableReference(this.context),
-      cursor: { element: this.element, nextSibling: null },
+      self: new UpdatableReference(context),
+      cursor: { element, nextSibling: null },
       dynamicScope: new TestDynamicScope(),
       mode: 'rehydrate'
     });
   }
 
-  renderTemplate(template: string): RenderResult {
-    this.template = template;
-    this.renderServerSide();
-    this.renderClientSide();
-    return this.renderResult!;
+  renderTemplate(template: string, context: Dict<Opaque>, element: HTMLElement, snapshot: () => void): RenderResult {
+    let serialized = this.renderServerSide(template, context, snapshot);
+    element.innerHTML = serialized;
+    return this.renderClientSide(template, context, element);
+  }
+
+  registerComponent(type: ComponentKind, name: string, layout: string): void {
+    registerComponent(this.clientEnv, type, name, layout);
+    registerComponent(this.serverEnv, type, name, layout);
+  }
+
+  registerHelper(name: string, helper: UserHelper): void {
+    this.clientEnv.registerHelper(name, helper);
+    this.serverEnv.registerHelper(name, helper);
   }
 }
 
@@ -747,13 +738,30 @@ export function test(...args: any[]) {
   return descriptor;
 }
 
-export function module(
+export interface RenderDelegateConstructor<Delegate extends RenderDelegate> {
+  new(): Delegate;
+}
+
+export interface RenderTestConstructor<D extends RenderDelegate, T extends AbstractRenderTest> {
+  new(delegate: D): T;
+}
+
+export function module<T extends AbstractRenderTest> (
   name: string,
-  klass: typeof AbstractRenderTest & Function,
+  klass: RenderTestConstructor<RenderDelegate, T>,
+  options = { componentModule: false }
+): void {
+  return rawModule(name, klass, TestEnvironmentRenderDelegate, options);
+}
+
+export function rawModule<D extends RenderDelegate, T extends AbstractRenderTest> (
+  name: string,
+  klass: RenderTestConstructor<D, T>,
+  Delegate: RenderDelegateConstructor<D>,
   options = { componentModule: false }
 ): void {
   if (options.componentModule) {
-    componentModule(name, klass);
+    componentModule(name, klass, Delegate);
   } else {
     QUnit.module(`[NEW] ${name}`);
 
@@ -761,7 +769,7 @@ export function module(
       const test = klass.prototype[prop];
 
       if (isTestFunction(test)) {
-        QUnit.test(prop, assert => test.call(new klass(), assert));
+        QUnit.test(prop, assert => test.call(new klass(new Delegate()), assert));
       }
     }
   }
@@ -775,10 +783,7 @@ interface ComponentTests {
   fragment: Function[];
 }
 
-function componentModule(
-  name: string,
-  klass: typeof AbstractRenderTest & Function
-) {
+function componentModule<D extends RenderDelegate, T extends AbstractRenderTest>(name: string, klass: RenderTestConstructor<D, T>, Delegate: RenderDelegateConstructor<D>) {
   let tests: ComponentTests = {
     glimmer: [],
     curly: [],
@@ -792,9 +797,9 @@ function componentModule(
     if (skip === true || test.skip === true) {
       shouldSkip = true;
     }
-    return (type: ComponentKind, klass: typeof AbstractRenderTest & Function) => {
-      let instance = new klass();
-      instance.testType = type;
+    return (type: ComponentKind, klass: RenderTestConstructor<D, T>) => {
+      let instance = new klass(new Delegate());
+      instance['testType'] = type;
       if (shouldSkip) {
         QUnit.skip(`${type.toLowerCase()}: ${prop}`, assert => test.call(instance, assert));
       } else {
