+++ packages/@glimmer/test-helpers/test/i-n-u-r-test.ts
@@ -1,15 +1,16 @@
-import { RenderTests } from "../lib/abstract-test-case";
+import { AbstractRenderTest } from "../lib/abstract-test-case";
+import { TestEnvironmentRenderDelegate } from "@glimmer/test-helpers";
 
 QUnit.module("Render Tests: I-N-U-R");
 
 QUnit.test("Can set properties", assert => {
-  new class extends RenderTests {
-    constructor() {
-      super();
+  new class extends AbstractRenderTest {
+    constructor(delegate: TestEnvironmentRenderDelegate) {
+      super(delegate);
       this.setProperties({ foo: "bar" });
       assert.equal(this.context.foo, "bar");
     }
-  }();
+  }(new TestEnvironmentRenderDelegate);
 });
 
 QUnit.test("Can take basic snapshots", assert => {
@@ -17,14 +18,14 @@ QUnit.test("Can take basic snapshots", assert => {
   let text = document.createTextNode("Foo");
   div.appendChild(text);
 
-  new class extends RenderTests {
+  new class extends AbstractRenderTest {
     element = div;
-    constructor() {
-      super();
+    constructor(delegate: TestEnvironmentRenderDelegate) {
+      super(delegate);
       let snapShot = this.takeSnapshot();
       assert.deepEqual(snapShot, [text, "up"]);
     }
-  }();
+  }(new TestEnvironmentRenderDelegate());
 });
 
 QUnit.test("Can take nested snapshots", assert => {
@@ -34,14 +35,14 @@ QUnit.test("Can take nested snapshots", assert => {
   p.appendChild(text);
   div.appendChild(p);
 
-  new class extends RenderTests {
+  new class extends AbstractRenderTest {
     element = div;
-    constructor() {
-      super();
+    constructor(delegate: TestEnvironmentRenderDelegate) {
+      super(delegate);
       let snapShot = this.takeSnapshot();
       assert.deepEqual(snapShot, [p, "down", text, "up", "up"]);
     }
-  }();
+  }(new TestEnvironmentRenderDelegate());
 });
 
 QUnit.test("Can take nested snapshots of serialized blocks", assert => {
@@ -53,12 +54,12 @@ QUnit.test("Can take nested snapshots of serialized blocks", assert => {
   div.appendChild(text);
   div.appendChild(close);
 
-  new class extends RenderTests {
+  new class extends AbstractRenderTest {
     element = div;
-    constructor() {
-      super();
+    constructor(delegate: TestEnvironmentRenderDelegate) {
+      super(delegate);
       let snapShot = this.takeSnapshot();
       assert.deepEqual(snapShot, [open, text, close, "up"]);
     }
-  }();
+  }(new TestEnvironmentRenderDelegate());
 });
